--- src/utils/addonManager.ts
@@ -1,8 +1,0 @@
-
-
-export class AddonManager{
-
-
-
-
-}
